+++ app/Http/Controllers/LoginController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use Framework\Authentication\Auth;
-use Framework\Database\Database;
 use Framework\Facades\Http;
 use Framework\Routing\BaseController;
 use Framework\Routing\ControllerInterface;
@@ -12,18 +12,17 @@ class LoginController extends BaseController implements ControllerInterface
 {
     public function execute(): void
     {
-        if ($this->getRequestMethod() === 'POST') {
+        if ($this->getRequestMethod() !== 'POST') {
             // TODO check if method is post -> else deny
+            Http::redirect('/');
         }
 
-        Database::prepared(
-            'SELECT `password` FROM users WHERE username = ?',
-            's',
-            $this->getParam('user_login')
-        );
-        
-        // TODO check password
-        $_SESSION['userId'] = 123; // TODO use user id
+        if (!Auth::isPasswordValid($this->getParam('username'), $this->getParam('password'))) {
+            // TODO Message that username or password is invalid
+            Http::redirect('/');
+        }
+
+        $_SESSION['userId'] = User::findByUsername($this->getParam('username'))->getId();
         Http::redirect('/');
     }
 }