+++ resources/database/migrations/2023_12_29_110250_create_products_table.php
@@ -14,7 +14,8 @@ return new class extends Migration
             'isDiscontinued tinyint(1) NOT NULL DEFAULT 0,' .
             'createdAt TIMESTAMP DEFAULT CURRENT_TIMESTAMP,' .
             'updatedAt DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,' .
-            'PRIMARY KEY (id)' .
+            'PRIMARY KEY (id),' .
+            'UNIQUE KEY `ukProductName` (name)' .
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;'
         );
     }