+++ resources/views/login.php
@@ -13,11 +13,11 @@ if (Auth::isLoggedIn()) {
 <?= component('layout.header') ?>
 
 <form action="login" method="post">
-    <label for="user_login" class="required"><?= __('Username') ?>:</label><br>
-    <input id="user_login" name="user_login" type="text" size="40" maxlength="250" required autofocus><br>
+    <label for="username" class="required"><?= __('Username') ?>:</label><br>
+    <input id="username" name="username" type="text" size="40" maxlength="250" required autofocus><br>
     
-    <label for="user_password" class="required"><?= __('Password') ?>:</label><br>
-    <input id="user_password" name="user_password" type="password" size="40"  maxlength="250" required><br>
+    <label for="password" class="required"><?= __('Password') ?>:</label><br>
+    <input id="password" name="password" type="password" size="40"  maxlength="250" required><br>
     <button><?= __('Login') ?></button>
 </form>